Build Suicide Safe eblast based on the bulletin
Changes I've made:
1. fixed HTML errors
2. inserted content from the .rtf document to the template
3. Please note: I kept section "Participate on the SAMHSA DTAC Discussion Board", "Receive The Dialogue", and "Questions About the SAMHSA DTAC" so that customers would have access to these resources. I did some research and found a previous bulletin containing these sections, and decided to keep them in the eblast.

Also, the link "Tell me when Suicide app is available" displays "invalid reference" in the rtf document. When doing this task, I was thinking like sending an real eblast to customers. Accordingly, I tried to find a valid link so that it can lead interested customers to SAMHSA website. So I added [https://public.govdelivery.com/accounts/USSAMHSA/subscriber/new], which I think is proper and that will lead users to the "sign up for SAMHSA e-mail updates" page to that sentence.